update Encapsulation into util packages #829
