+++ pkg/datasource/sql/undo/builder/basic_undo_log_builder.go
@@ -26,6 +26,7 @@ import (
 	"github.com/arana-db/parser/test_driver"
 	gxsort "github.com/dubbogo/gost/sort"
 	"io"
+	"seata.apache.org/seata-go/pkg/datasource/sql/util"
 	"strings"
 
 	"seata.apache.org/seata-go/pkg/datasource/sql/types"
@@ -275,48 +276,5 @@ func (b *BasicUndoLogBuilder) buildLockKey(rows driver.Rows, meta types.TableMet
 
 // the string as local key. the local key example(multi pk): "t_user:1_a,2_b"
 func (b *BasicUndoLogBuilder) buildLockKey2(records *types.RecordImage, meta types.TableMeta) string {
-	var lockKeys strings.Builder
-	type ColMapItem struct {
-		pkIndex  int
-		colIndex int
-	}
-
-	lockKeys.WriteString(meta.TableName)
-	lockKeys.WriteString(":")
-
-	keys := meta.GetPrimaryKeyOnlyName()
-	keyIndexMap := make(map[string]int, len(keys))
-	for idx, columnName := range keys {
-		keyIndexMap[columnName] = idx
-	}
-
-	columns := make([]ColMapItem, 0, len(keys))
-	if len(records.Rows) > 0 {
-		for colIdx, column := range records.Rows[0].Columns {
-			if pkIdx, ok := keyIndexMap[column.ColumnName]; ok {
-				columns = append(columns, ColMapItem{pkIndex: pkIdx, colIndex: colIdx})
-			}
-		}
-		for i, row := range records.Rows {
-			if i > 0 {
-				lockKeys.WriteString(",")
-			}
-			primaryKeyValues := make([]interface{}, len(keys))
-			for _, mp := range columns {
-				if mp.colIndex < len(row.Columns) {
-					primaryKeyValues[mp.pkIndex] = row.Columns[mp.colIndex].Value
-				}
-			}
-			for j, pkVal := range primaryKeyValues {
-				if j > 0 {
-					lockKeys.WriteString("_")
-				}
-				if pkVal == nil {
-					continue
-				}
-				lockKeys.WriteString(fmt.Sprintf("%v", pkVal))
-			}
-		}
-	}
-	return lockKeys.String()
+	return util.BuildLockKey(records, meta)
 }
